fix(packages): fix offline enterprise server package config (#303)
Before this, I ideally thought that there was no difference in the file
list contained in the tiup package in the tiup offline package, but only
the difference in the program content of the two processes. But in fact,
due to the lack of transparency and precision in many historical
processes, there is a direct and undue difference between the enterprise
version and the community version.

For continued consistency, the configuration is updated here to maintain
the status quo. But here we use precise file names to control, rather
than directories.

Signed-off-by: wuhuizuo <wuhuizuo@126.com>

---------

Signed-off-by: wuhuizuo <wuhuizuo@126.com>
Co-authored-by: Purelind <purelind@users.noreply.github.com>