update image of tidb operator v2 (#620)
This pull request updates the naming conventions for container images
and their corresponding build stages in the
`packages/packages.yaml.tmpl` file to align with the `tidb-operator`
project. The changes improve clarity and consistency in the artifact
definitions.

### Naming updates for container images and build stages:
* Renamed the container image `operator` to `tidb-operator` and updated
the corresponding build stage to `tidb-operator`.
* Renamed the container image `backup-manager` to `tidb-backup-manager`
and updated the corresponding build stage to `tidb-backup-manager`.


Ref https://github.com/pingcap/tidb-operator/pull/6200